# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A simple RESTful API server built with Go, Gin, sqlc, and golang-migrate.",
        "title": "Mini Lab API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.minilab.com/support",
            "email": "support@minilab.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/machine": {
            "get": {
                "description": "Get a list of all machines",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get all machines",
                "responses": {
                    "200": {
                        "description": "List of machines",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Machine"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new machine with optional task type assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Create a new machine",
                "parameters": [
                    {
                        "description": "Machine creation data",
                        "name": "machine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Machine created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/machine/{id}": {
            "get": {
                "description": "Get a single machine by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get machine by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid machine ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing machine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Update a machine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Machine update data",
                        "name": "machine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a machine by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Delete a machine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid machine ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/roles": {
            "get": {
                "description": "Get a list of all available roles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get all roles",
                "responses": {
                    "200": {
                        "description": "List of roles",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Role"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "description": "Get a single role by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get role by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Role"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid role ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/task_type": {
            "get": {
                "description": "Get a list of all task types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-types"
                ],
                "summary": "Get all task types",
                "responses": {
                    "200": {
                        "description": "List of task types",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.TaskType"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new task type with optional machine assignments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-types"
                ],
                "summary": "Create a new task type",
                "parameters": [
                    {
                        "description": "Task type creation data",
                        "name": "task_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task type created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TaskType"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/task_type/{id}": {
            "get": {
                "description": "Get a single task type by its ID with associated machines",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-types"
                ],
                "summary": "Get task type by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task type details with machines",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TaskType"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid task type ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Task type not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing task type by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-types"
                ],
                "summary": "Update a task type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task type update data",
                        "name": "task_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTaskTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task type updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TaskType"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Task type not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task type by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-types"
                ],
                "summary": "Delete a task type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task type deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid task type ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Task type not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "put": {
                "description": "Assign task types to a user (only for members)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Assign task types to a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task type assignment data",
                        "name": "task_types",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserTaskAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task types assigned successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data or user role",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Error message"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.CreateMachineRequest": {
            "type": "object",
            "required": [
                "estimate_time",
                "name",
                "quantity"
            ],
            "properties": {
                "estimate_time": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 120
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model X"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "task_type": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CreateTaskTypeRequest": {
            "type": "object"
        },
...
        "models.Machine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "estimate_time": {
                    "description": "in minutes",
                    "type": "integer",
                    "example": 120
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model X"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "task_type": {
                    "type": "string",
                    "example": "3D Printing"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "models.UpdateMachineRequest": {
            "type": "object",
            "properties": {
                "estimate_time": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 90
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model Y"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 3
                }
            }
        },
        "models.UpdateTaskTypeRequest": {
            "type": "object"
        },
...
        "models.UserTaskAssignmentRequest": {
            "type": "object"
        }
    }
}